# docs/openapi.yaml
openapi: 3.0.3
info:
  title: QuickQuiz Dev API
  version: 0.1.0
  description: QuickQuiz Dev MVP API.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/session/reset:
    post:
      tags:
        - Session
      summary: Resets the current session.
      description: Removes in-memory runs associated with the provided session header.
      operationId: resetSession
      parameters:
        - $ref: "#/components/parameters/Theme"
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: Session reset.
          content:
            application/json:
              schema:
                type: object
                required:
                  - reset
                properties:
                  reset:
                    type: boolean
                    example: true
        "400":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/runs:
    post:
      tags:
        - Runs
      summary: Creates a new run.
      operationId: createRun
      parameters:
        - $ref: "#/components/parameters/Theme"
        - $ref: "#/components/parameters/SessionID"
        - $ref: "#/components/parameters/Locale"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: Run created with the first question.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRunResponse"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "409":
          $ref: "#/components/responses/NoQuestionsLeft"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    Theme:
      name: X-QuickQuiz-Theme
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        example: dev
      description: Active quiz theme requested by the frontend distribution.
    Locale:
      name: X-QuickQuiz-Locale
      in: header
      required: false
      schema:
        type: string
        example: en-US
      description: BCP 47 content locale. Explicit locale takes precedence over Accept-Language and fallback locale.
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        example: pt-BR,pt;q=0.9,en-US;q=0.8
      description: Browser/client locale preference used only when no explicit locale is provided.
    SessionID:
      name: X-QuickQuiz-Session-ID
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      example: 8dc52f2f-ecbc-448c-a10a-684297c56d8f
...
  schemas:
    Difficulty:
      type: integer
      format: int32
      enum:
        - 1
        - 2
        - 3
        - 4
      description: "1=easy, 2=normal, 3=hard, 4=hardcore."
...
    CreateRunRequest:
      type: object
      required:
        - topic
        - difficulty
      properties:
        locale:
          type: string
          example: pt-BR
        topic:
          type: string
          example: go
        difficulty:
          $ref: "#/components/schemas/Difficulty"
    CreateRunResponse:
      type: object
      required:
        - runId
        - theme
        - locale
        - question
      properties:
        runId:
          type: string
          example: run_014c27d8d07ed827411d70c1edfdcd96
        theme:
          type: string
          example: dev
        locale:
          type: string
          example: pt-BR
        question:
          $ref: "#/components/schemas/PublicQuestion"
    PublicQuestion:
      type: object
      required:
        - id
        - prompt
        - options
        - current
        - total
      properties:
        id:
          type: string
          example: go-easy-001
        prompt:
          type: string
          example: Qual comando executa os testes de um modulo Go?
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizOption"
        current:
          type: integer
          example: 1
        total:
          type: integer
          example: 10
    QuizOption:
      type: object
      required:
        - id
        - text
      properties:
        id:
          type: string
          example: opt_4a5f5a1a96adac6b
        text:
          type: string
          example: go test ./...
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: invalid_input
            message:
              type: string
              example: Invalid input
  responses:
    InvalidInput:
      description: Invalid input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NoQuestionsLeft:
      description: No questions available.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"